fix(react): bundle bridge and scalars
Into single file packages to avoid strange import
issues with esbuild.
