--- packages/npm/@amazeelabs/bridge/src/index.ts
@@ -1,1 +1,0 @@
-export { Link, navigate, useLocation } from './link';
